KOALA v2.1.0 - CHANGELOG
 Koala is now COMPATIBLE WITH KONTAKT 5.6. Please read this Changelog
carefully!

* Kontakt 5.6.5 support: XY Pad and new Math runtime
* New UI functions and color/status constants
* New function to load all the UI Ids of a Family inside an array
* New set of math functions with a killer curve function
* New integer conversion functions (inspired by the majestic work of
BigBob † and his Math library)
* XY Pad is now fully supported by Koala and the Debugger: advanced
functions to link cursors to UI controls or Engine Pars
* new automation functions
* new functions to convert the value of a UI control into a percentage
value (0 to 100) or a decimal value (0 to 1, real number)
* Changed responsivity of the Debugger according to the UI size
* Koala is backward compatible with Kontakt 5.5.2, but the functions
that use real numbers will return an error when applying the script to
Kontakt 5.5.2. The same goes for XY pad, which has been introduced with
Kontakt 5.6.5, thus it is not available in prior versions.

* Fixed create_text_edit macro
* Fixed a bug in the Logger which prevented the Persistence callback to
be logged properly. The Logger cannot be used in the on init callback
* Improved 'timer' function. Now the data from 'timer' are loaded into
a text variable in order to save a lot of scripting lines. New usage:
	call timer
	message(@timer)